refactor: simplify pickups across PR #1201 (#1188 follow-on for shared helpers)
Three parallel review agents (reuse / quality / efficiency) all
converged on the same handful of cleanups. Applying the
high-value ones:

## 1. Hoist `identifierPath` + `stringifyParsedExpr` to @barefootjs/jsx

The Mojolicious adapter open-coded both helpers verbatim from the
Go adapter (#1188). With a third adapter likely to follow, the
duplication was about to compound. Both are pure ParsedExpr
utilities — moved to `expression-parser.ts` and exported from the
package entry. Mojo and Go now import them; their local copies
are gone.

`stringifyParsedExpr` is a NEW export distinct from the existing
`exprToString`. The two have different goals: `exprToString` is a
debug formatter (`[UNSUPPORTED: …]`, bare string-literal quoting);
`stringifyParsedExpr` is a round-trip-safe re-stringifier
(`JSON.stringify` for string literals, computed-member quoting,
raw passthrough for unsupported). Adapters that re-feed AST
output back into a downstream pipeline want the latter.

## 2. Drop the parallel `templatePrimitiveArities` map

The earlier shape had two parallel maps derived from a single
`MOJO_TEMPLATE_PRIMITIVES` spec, plus a regression test pinning
their sync. The arity gate in `rewriteTemplatePrimitives` now
reads the spec record directly (`MOJO_TEMPLATE_PRIMITIVES[path].arity`)
— one less moving part, the sync invariant becomes a
type-checker invariant rather than a runtime test.

## 3. Add substring pre-check before `parseExpression`

`convertExpressionToPerl` runs on every JSX expression in the
compile pipeline. The new pre-pass added a `parseExpression` call
(which invokes `ts.createSourceFile` — the dominant cost). For
the common case of "no primitive callee in the source string", a
cheap `RegExp.test` against the registered keys short-circuits
before the parser runs. The regex is built once at module load
from the registry keys, so it stays in sync as primitives are
added.

## 4. Module-scope the `templatePrimitives` Map

`templatePrimitives: TemplatePrimitiveRegistry = Object.fromEntries(...)`
as a class field rebuilt the same Map on every `new MojoAdapter()`.
Lifted to module scope (`MOJO_PRIMITIVE_EMIT_MAP`) since it's the
same for every instance.

## 5. Real numeric NaN sentinel in BarefootJS.pm

The Perl helpers used the literal string `"NaN"` as a sentinel
for parse failure. That conflated two different cases: (a) a
user passing the actual string `"NaN"`, which `looks_like_number`
accepts and converts to a real numeric NaN, vs (b) a parse
failure, which returned the literal string. Downstream `eq 'NaN'`
checks would then disagree depending on which path produced the
value.

Switched to real numeric NaN via `0 + 'nan'` and a portable
`$x != $x` detection (NaN is the only float that isn't equal to
itself). The sentinel is now what JS callers expect:
`Math.floor(NaN) === NaN`, propagating through arithmetic
without special-casing strings.

Tests updated to use a local `is_nan` helper rather than string
comparison (NaN stringifies as "NaN" / "nan" platform-dependently).

## Drive-by

- `mojo-adapter.test.ts`'s "two-tier source-of-truth keeps emit +
  arity in sync" test was pinning a derivation to itself; with
  the parallel map gone the test has nothing to verify.
- The Go adapter (`packages/adapter-go-template/...`) imports
  `identifierPath` from `@barefootjs/jsx` now and drops its own
  module-private copy. Go's other helpers (the spec record, the
  regex pre-check pattern) could follow in a separate PR — out
  of scope here.

Test plan:
- jsx: 1008 pass — no regressions
- adapter-hono: 96 pass — no regressions
- adapter-go-template: 81 pass — local `identifierPath` removal
  proven by the existing test surface
- adapter-mojolicious: 62 pass / 2 skip / 0 fail (was 63 — the
  removed sync test accounts for the −1)
