+++ packages/jsx/src/index.ts
@@ -210,7 +210,7 @@ export {
 export { ErrorCodes, createError, formatError, generateCodeFrame } from './errors'
 
 // Expression Parser
-export { parseExpression, isSupported, exprToString, parseBlockBody } from './expression-parser'
+export { parseExpression, isSupported, exprToString, stringifyParsedExpr, identifierPath, parseBlockBody } from './expression-parser'
 export type { ParsedExpr, ParsedStatement, SupportLevel, SupportResult, TemplatePart } from './expression-parser'
 
 // Debug analysis
